backend/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==3.2.2
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
jax==0.8.1
jaxlib==0.8.1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
proto-plus==1.27.0
protobuf==4.25.3
psutil==7.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.4
SQLAlchemy==2.0.44
srt==3.5.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.0
uvicorn==0.38.0
vosk==0.3.45
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
